feat(granite): adiciona Granite 4.1 NAR e corrige convencao da attention_mask (1=real, 0=pad)

+++ tools/granite/export_onnx.py
@@ -92,7 +92,7 @@ def main():
         ["logits"],
         {
             "input_features": feats["input_features"].numpy(),
-            "attention_mask": feats["attention_mask"].numpy().astype(np.float32),
+            "attention_mask": feats["attention_mask"].numpy().astype(np.int64),
         },
     )[0]
     ids_onnx = logits_onnx.argmax(axis=-1)[0].tolist()
